API

Contents

Index

GnssDates.GnssDates — Module

Package with definitions to work with GNSS datetimes and time intervals.

Abstract types: SystemTime, CoarseTime, FineTime, TimeDelta

Time references: GnssTime, GPST, GST

Time intervals: CoarseTimeDelta, FineTimeDelta

Operators: +, -, ==, >, >=, <, <=

GnssDates.GAL_WEEK_OFFSET — Constant

Week Number offset between GST and GnssTime.

GnssDates.GPST₀ — Constant

Origin of GPST.

GnssDates.GST₀ — Constant

Origin of GST.

GnssDates.LEAP_SECOND_TAI_OFFSET — Constant

Time difference in seconds between UTC and TAI at origin of GnssTime.

GnssDates.SECONDS_IN_WEEK — Constant

Number of seconds in a week.

GnssDates.CoarseTime — Type

Abstract supertype for all coarse time references.

Coarse time references are characterized for always having two fields:

wn::Int64: integer weeks elapsed since origin of time reference.

tow::Int64: integer seconds elapsed since beginning of week.

GnssDates.CoarseTimeDelta — Type

CoarseTimeDelta(weeks, seconds)

Time interval with integer-second resolution.

The constructor can convert automatically from other date and datetime types.

GnssDates.FineTime — Type

Abstract supertype for all fine time references.

Fine time references are characterized for always having three fields:

wn::Int64: integer weeks elapsed since origin of time reference.

tow_int::Int64: integer seconds elapsed since beginning of week.

tow_frac::Int64: seconds elapsed since tow_int as a floating point number.

Currently, the only implementation of a FineTime is GnssTime.

GnssDates.FineTimeDelta — Type

FineTimeDelta(weeks, seconds, seconds_frac)

Time interval with subsecond resolution.

The constructor can convert automatically from other date and datetime types.

GnssDates.GPST — Type

GPST(wn, tow) <: CoarseTime

Coarse time reference as disseminated by GPS.

Resolution

It has integer-second resolution because TOW is represented by an Int64. WN does not rollover (it keeps counting since GPST₀ continuously.)

More info

Check Time References in GNSS for more info.

GnssDates.GPST — Method

GPST(t::DateTime, UTC)

Convert from DateTime assumming that t is a UTC time.

GnssDates.GPST — Method

GPST(t::T)

Convert from time reference of type T to GPST.

Valid types are:

T::SystemTime

T::Date

T::DateTime

GnssDates.GST — Type

GST(wn, tow) <: CoarseTime

Coarse time reference as disseminated by Galileo.

Resolution

It has integer-second resolution because TOW is represented by an Int64. WN does not rollover (it keeps counting since GST₀ continuously.)

More info

Check Time References in GNSS for more info.

GnssDates.GST — Method

GST(t::DateTime, UTC)

Convert from DateTime assumming that t is a UTC time.

GnssDates.GST — Method

GST(t::T)

Convert from time reference of type T to GST.

Valid types are:

T::SystemTime

T::Date

T::DateTime

GnssDates.GnssTime — Type

GnssTime(wn, tow_int, tow_frac) <: FineTime

Absolute time reference whose origin is aligned to GPS's.

Resolution

It has subsecond resolution down to femtoseconds because it uses Float64 to represent decimal part of the current second. WN does not rollover (it keeps counting since GPST₀ continuously.)

GnssDates.GnssTime — Method

GnssTime(t::DateTime, UTC)

Convert from DateTime assumming that t is a UTC time.

GnssDates.GnssTime — Method

GnssTime(t::T)

Convert from time reference of type T to GnssTime.

Valid types are:

T::SystemTime

T::Date

T::DateTime

GnssDates.SystemTime — Type

Abstract supertype for all time references.

GnssDates.TimeDelta — Type

Abstract supertype for all time intervals.

GnssDates._canon_coarsedelta — Method

_canon_coarsedelta(w, s) -> Tuple{Int,Int}

Normalize a (w, s) tuple following that:

s ∈ [0, SECONDS_IN_WEEK) if it is a positive interval

s ∈ [0, -SECONDS_IN_WEEK) if it is a negative interval

GnssDates._canon_coarsetime — Method

_canon_coarsetime(wn, tow) -> Tuple{Int,Int}

Normalize a (wn, tow) tuple following that:

tow ∈ [0, SECONDS_IN_WEEK)

GnssDates._canon_finedelta — Method

_canon_finedelta(w, s, sf) -> Tuple{Int,Int}

Normalize a (w, s, sf) tuple following that:

s ∈ [0, SECONDS_IN_WEEK) if it is a positive interval

s ∈ [0, -SECONDS_IN_WEEK) if it is a negative interval

sf ∈ [0.0, 1.0) if it is a positive interval

s ∈ [0, -1.0) if it is a negative interval

GnssDates._canon_finetime — Method

_canon_finetime(wn, tow, towf) -> Tuple{Int,Int,Float64}

Normalize a (wn, tow, towf) tuple following that:

tow ∈ [0, SECONDS_IN_WEEK)

towf ∈ [0.0, 1.0)